LPD-23693 Rename packages
