+++ modules/apps/site-initializer/site-initializer-raylife-ap/src/main/java/com/liferay/site/initializer/raylife/ap/internal/kaleo/runtime/action/executor/ReviewedActionExecutor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package site.initializer.workflow.definitions.underwriting.action.executor;
+package com.liferay.site.initializer.raylife.ap.internal.kaleo.runtime.action.executor;
 
 import com.liferay.portal.workflow.kaleo.model.KaleoAction;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
